requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
python-dotenv
django-unfold
drf-spectacular[sidecar]
django-cors-headers
